requirements.txt
click==6.6
cov-core==1.15.0
coverage==3.7.1
Flask==0.11.1
Flask-SQLAlchemy==2.1
itsdangerous==0.24
Jinja2==2.8
lxml==3.6.4
MarkupSafe==0.23
mock==2.0.0
pbr==1.10.0
Pillow==3.4.2
py==1.4.31
pyjwt==1.4.2
pytest==3.0.3
pytest-cov==2.4.0
PyYAML==3.12
requests==2.11.1
sentinels==1.0.0
six==1.9.0
SQLAlchemy==1.1.4
virtualenv==15.0.3
Werkzeug==0.11.11
